Refector preview image discovery into shared location
