+++ src/project/types/website/listing/website-listing.ts
@@ -26,6 +26,7 @@ import {
 import { renderEjs } from "../../../../core/ejs.ts";
 import { resourcePath } from "../../../../core/resources.ts";
 import { lines } from "../../../../core/text.ts";
+import { findDescriptionMd, findPreviewImgMd } from "../util/discover-meta.ts";
 
 // The core listing type
 export interface Listing {
@@ -97,9 +98,9 @@ export async function listingHtmlDependencies(
       // Create the item
       const documentMeta = target?.markdown.yaml;
       const description = documentMeta?.description as string ||
-        defaultDescription(target?.markdown.markdown);
+        findDescriptionMd(target?.markdown.markdown);
       const image = documentMeta?.image as string ||
-        defaultImage(target?.markdown.markdown);
+        findPreviewImgMd(target?.markdown.markdown);
       const date = documentMeta?.date as Date;
       const author = Array.isArray(documentMeta?.author)
         ? documentMeta?.author
@@ -403,59 +404,4 @@ function resolveListingStr(val: string): Listing {
     contents: [val],
     classes: [],
   };
-}
-
-const previewClassImg =
-  /!\[[^\]]*\]\((.*?)(?:\".*\")?\)\{[^\|]*\.quarto-preview[\s\}]+/;
-const markdownImg = /!\[[^\]]*\]\((.*?)(?:\".*\")?\)(?:\{(?:[^\|]*)\})?/;
-
-function defaultImage(markdown?: string): string | undefined {
-  if (markdown) {
-    // Look for an explictly tagged image
-    const explicitMatch = markdown.match(previewClassImg);
-    if (explicitMatch) {
-      return explicitMatch[1];
-    }
-
-    // Otherwise select the first image
-    const match = markdown.match(markdownImg);
-    if (match) {
-      return match[1];
-    }
-  }
-  return undefined;
-}
-
-function defaultDescription(markdown?: string): string | undefined {
-  if (markdown) {
-    const previewText: string[] = [];
-    let accum = false;
-
-    // Controls what counts as ignorable lines (empty or markdown of
-    // specific types)
-    const skipLines = [/^\#+/, /^\:\:\:[\:]*/];
-    const emptyLine = (line: string) => {
-      return line.trim() === "";
-    };
-
-    // Go through each line and find the first paragraph, then accumulate
-    // that text as the description
-    for (const line of lines(markdown)) {
-      if (!accum) {
-        // When we encounter the first
-        if (!emptyLine(line) && !skipLines.find((skip) => line.match(skip))) {
-          accum = true;
-          previewText.push(line);
-        }
-      } else {
-        if (emptyLine(line) || skipLines.find((skip) => line.match(skip))) {
-          break;
-        } else {
-          previewText.push(line);
-        }
-      }
-    }
-    return previewText.join("\n");
-  }
-  return undefined;
 }
